change id from int to long

+++ src/main/java/com/iam/user/repository/UserRepository.java
@@ -8,7 +8,7 @@ import java.util.Optional;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
-    Optional<User> findById(Integer userId);
+    Optional<User> findById(Long userId);
 
     Optional<User> findByEmail(String email);
